Fixes bug: Biomart returning incorrect assembly (hg38)
The BioMart host migration in January 2016 required us to change the URL
for our gene query services. To correct this it was changed to the
latest Ensembl release. This release however was returning hg38 assembly
locations instead of hg19.

Switching back to the previous URL appears to, again, work. In the
future we should find a suitable backup URL which returns hg19 assembly
genomic locations.