Implements quick delete of projects for CLI use
New method to delete projects which bypasses Hibernate cascade
management. This is so that Hibernate does not unnecessarily load all
project related objects before deletion. Shows a improvement in deletion
time for large (special) projects. 

In order to do this the Characteristic Entity had to be made
bidirectional with Variant (also removes related join table).