Fix genes associated with phenotype dialog
Requires replacing find-phenotype-genes Gemma endpoint with
find-candidate-genes and manually choosing a single
(most specific) term.

This currently does not attempt to find which of the given
terms are msot specific and instead chooses the first entry.

The reasoning behind this is that the better option would
be the change the front-end to better be able to display
all associated terms rather than just the most specific.
As the most specific term is a vague concept when you
have children of separate branches.